+++ test/OpenTelemetry.Tests/Internal/GuardTest.cs
@@ -15,6 +15,7 @@
 // </copyright>
 
 using System;
+using System.Runtime.CompilerServices;
 using System.Threading;
 using Xunit;
 
@@ -172,4 +173,34 @@ namespace OpenTelemetry.Internal.Tests
             Assert.Equal("0", ex1.ParamName);
         }
     }
+
+#if !NETCOREAPP3_0_OR_GREATER
+#pragma warning disable SA1402 // File may only contain a single type
+    /// <summary>
+    /// Borrowed from: <see href="https://github.com/dotnet/runtime/blob/main/src/libraries/System.Runtime/tests/System/Runtime/CompilerServices/CallerArgumentExpressionAttributeTests.cs"/>.
+    /// </summary>
+    public class CallerArgumentExpressionAttributeTests
+    {
+        [Theory]
+        [InlineData(null)]
+        [InlineData("")]
+        [InlineData("paramName")]
+        public static void Ctor_ParameterName_Roundtrip(string value)
+        {
+            var caea = new CallerArgumentExpressionAttribute(value);
+            Assert.Equal(value, caea.ParameterName);
+        }
+
+        [Fact]
+        public static void BasicTest()
+        {
+            Assert.Equal("\"hello\"", GetValue("hello"));
+            Assert.Equal("3 + 2", GetValue(3 + 2));
+            Assert.Equal("new object()", GetValue(new object()));
+        }
+
+        private static string GetValue(object argument, [CallerArgumentExpression("argument")] string expr = null) => expr;
+    }
+#pragma warning restore SA1402 // File may only contain a single type
+#endif
 }
